renamed backend_output node to robot_communication
