+++ src/thunderbots/software/launch/ai.launch
@@ -8,8 +8,8 @@
     <node name="ai_logic" pkg="thunderbots" type="ai_logic" output="screen">
     </node>
 
-    <!-- Launch the backend_output node -->
-    <node name="backend_output" pkg="thunderbots" type="backend_output" output="screen">
+    <!-- Launch the robot_communication node -->
+    <node name="robot_communication" pkg="thunderbots" type="robot_communication" output="screen">
     </node>
 
 </launch>
